+++ src/nix/app.cc
@@ -79,9 +79,19 @@ UnresolvedApp Installable::toApp(EvalState & state)
     if (type == "app") {
         auto [program, context] = cursor->getAttr("program")->getStringWithContext();
 
-        std::vector<StorePathWithOutputs> context2;
-        for (auto & [path, name] : context)
-            context2.push_back({path, {name}});
+        std::vector<DerivedPath> context2;
+        for (auto & [path, name] : context) {
+            context2.push_back(name != "" || path.isDerivation()
+                ? (DerivedPath) DerivedPath::Built {
+                    .drvPath = path,
+                    .outputs = name != ""
+                        ? StringSet { name }
+                        : StringSet { },
+                }
+                : (DerivedPath) DerivedPath::Opaque {
+                    .path = path,
+                });
+        }
 
         return UnresolvedApp{App {
             .context = std::move(context2),
@@ -105,7 +115,10 @@ UnresolvedApp Installable::toApp(EvalState & state)
             : DrvName(name).name;
         auto program = outPath + "/bin/" + mainProgram;
         return UnresolvedApp { App {
-            .context = { { drvPath, {outputName} } },
+            .context = { DerivedPath::Built {
+                .drvPath = drvPath,
+                .outputs = {outputName},
+            } },
             .program = program,
         }};
     }
@@ -123,7 +136,7 @@ App UnresolvedApp::resolve(ref<Store> evalStore, ref<Store> store)
 
     for (auto & ctxElt : unresolved.context)
         installableContext.push_back(
-            std::make_shared<InstallableDerivedPath>(store, ctxElt.toDerivedPath()));
+            std::make_shared<InstallableDerivedPath>(store, ctxElt));
 
     auto builtContext = Installable::build(evalStore, store, Realise::Outputs, installableContext);
     res.program = resolveString(*store, unresolved.program, builtContext);
